#68 rename: auth 관련 에러 핸들러 패키지 위치 수정
