+++ src/main/java/com/carumuch/capstone/domain/auth/handler/CustomAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.carumuch.capstone.global.common.exception;
+package com.carumuch.capstone.domain.auth.handler;
 
 import com.carumuch.capstone.global.common.ResponseDto;
 import lombok.RequiredArgsConstructor;
@@ -25,7 +25,7 @@ public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint 
     @Override
     public void commence(HttpServletRequest request,
                          HttpServletResponse response,
-                         AuthenticationException authException) throws IOException, ServletException {
+                         AuthenticationException authException) throws IOException {
 
         response.setStatus(HttpStatus.UNAUTHORIZED.value());
         response.setCharacterEncoding("utf-8");
